combine goals files
