--- artifactory/healcomponents/maven/goals_parse_test.go
@@ -1,30 +1,0 @@
-package maven
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestParseMavenCLIArgs_FileFlag(t *testing.T) {
-	pom, projects := parseMavenCLIArgs([]string{"-f", "sub/pom.xml", "install"})
-	assert.Equal(t, "sub/pom.xml", pom)
-	assert.Nil(t, projects)
-
-	pom, _ = parseMavenCLIArgs([]string{"--file", "other/pom.xml", "clean", "install"})
-	assert.Equal(t, "other/pom.xml", pom)
-}
-
-func TestParseMavenCLIArgs_ProjectsFlag(t *testing.T) {
-	_, projects := parseMavenCLIArgs([]string{"-pl", "mod-a,mod-b", "install"})
-	assert.Equal(t, []string{"mod-a", "mod-b"}, projects)
-
-	_, projects = parseMavenCLIArgs([]string{"--projects", "services/api", "verify"})
-	assert.Equal(t, []string{"services/api"}, projects)
-}
-
-func TestParseMavenCLIArgs_NoFlags(t *testing.T) {
-	pom, projects := parseMavenCLIArgs([]string{"clean", "install"})
-	assert.Empty(t, pom)
-	assert.Nil(t, projects)
-}
